refactor: 여행 공유 관련 로직 Trip 패키지로 통합
