+++ backend/src/main/java/hanglog/trip/domain/repository/SharedTripRepository.java
@@ -1,6 +1,6 @@
-package hanglog.share.domain.repository;
+package hanglog.trip.domain.repository;
 
-import hanglog.share.domain.SharedTrip;
+import hanglog.trip.domain.SharedTrip;
 import java.util.List;
 import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
