+++ backend/src/main/java/hanglog/trip/dto/request/SharedStatusRequest.java
@@ -1,4 +1,4 @@
-package hanglog.share.dto.request;
+package hanglog.trip.dto.request;
 
 import static lombok.AccessLevel.PRIVATE;
 
@@ -10,7 +10,7 @@ import lombok.NoArgsConstructor;
 @Getter
 @AllArgsConstructor
 @NoArgsConstructor(access = PRIVATE)
-public class SharedTripStatusRequest {
+public class SharedStatusRequest {
 
     @NotNull(message = "공유 상태를 선택해주세요.")
     private Boolean sharedStatus;
