+++ backend/src/main/java/hanglog/trip/presentation/SharedTripController.java
@@ -1,21 +1,13 @@
-package hanglog.share.presentation;
+package hanglog.trip.presentation;
 
-import hanglog.auth.Auth;
-import hanglog.auth.domain.Accessor;
+import hanglog.trip.service.SharedTripService;
+import hanglog.trip.dto.response.TripDetailResponse;
 import hanglog.trip.dto.response.TripLedgerResponse;
 import hanglog.trip.service.LedgerService;
-import hanglog.share.dto.request.SharedTripStatusRequest;
-import hanglog.share.dto.response.SharedTripCodeResponse;
-import hanglog.share.service.SharedTripService;
-import hanglog.trip.dto.response.TripDetailResponse;
-import hanglog.trip.service.TripService;
-import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PatchMapping;
 import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RestController;
 
 
@@ -24,7 +16,6 @@ import org.springframework.web.bind.annotation.RestController;
 public class SharedTripController {
 
     private final SharedTripService sharedTripService;
-    private final TripService tripService;
     private final LedgerService ledgerService;
 
     @GetMapping("/shared-trips/{sharedCode}")
@@ -32,20 +23,6 @@ public class SharedTripController {
         final Long tripId = sharedTripService.getTripId(sharedCode);
         final TripDetailResponse tripDetailResponse = sharedTripService.getSharedTripDetail(tripId);
         return ResponseEntity.ok().body(tripDetailResponse);
-    }
-
-    @PatchMapping("/trips/{tripId}/share")
-    public ResponseEntity<SharedTripCodeResponse> updateSharedStatus(
-            @Auth final Accessor accessor,
-            @PathVariable final Long tripId,
-            @RequestBody @Valid final SharedTripStatusRequest sharedTripStatusRequest
-    ) {
-        tripService.validateTripByMember(accessor.getMemberId(), tripId);
-        final SharedTripCodeResponse sharedTripCodeResponse = sharedTripService.updateSharedTripStatus(
-                tripId,
-                sharedTripStatusRequest
-        );
-        return ResponseEntity.ok().body(sharedTripCodeResponse);
     }
 
     @GetMapping("/shared-trips/{sharedCode}/expense")
