+++ backend/src/test/java/hanglog/share/fixture/ShareFixture.java
@@ -3,7 +3,7 @@ package hanglog.share.fixture;
 import static hanglog.integration.IntegrationFixture.MEMBER;
 
 import hanglog.city.domain.City;
-import hanglog.share.domain.SharedTrip;
+import hanglog.trip.domain.SharedTrip;
 import hanglog.trip.domain.DayLog;
 import hanglog.trip.domain.Trip;
 import hanglog.trip.domain.type.PublishedStatusType;
