+++ backend/src/test/java/hanglog/trip/presentation/TripControllerTest.java
@@ -42,6 +42,8 @@ import hanglog.auth.domain.MemberTokens;
 import hanglog.city.domain.City;
 import hanglog.expense.domain.CategoryExpense;
 import hanglog.expense.domain.DayLogExpense;
+import hanglog.trip.dto.request.SharedStatusRequest;
+import hanglog.trip.dto.response.SharedCodeResponse;
 import hanglog.trip.dto.response.TripLedgerResponse;
 import hanglog.global.ControllerTest;
 import hanglog.trip.dto.request.PublishedStatusRequest;
@@ -728,6 +730,72 @@ class TripControllerTest extends ControllerTest {
                                         .description("여행 ID")
                         )
                 ));
+    }
+
+    @DisplayName("공유 상태를 변경한다")
+    @Test
+    void updateSharedStatus() throws Exception {
+        // given
+        final SharedStatusRequest sharedStatusRequest = new SharedStatusRequest(true);
+        final SharedCodeResponse sharedCodeResponse = new SharedCodeResponse("sharedCode");
+        when(tripService.updateSharedTripStatus(anyLong(), any(SharedStatusRequest.class)))
+                .thenReturn(sharedCodeResponse);
+        given(refreshTokenRepository.existsByToken(any())).willReturn(true);
+        doNothing().when(jwtProvider).validateTokens(any());
+        given(jwtProvider.getSubject(any())).willReturn("1");
+        doNothing().when(tripService).validateTripByMember(anyLong(), anyLong());
+
+        // when & then
+        mockMvc.perform(patch("/trips/{tripId}/share", 1)
+                        .header(AUTHORIZATION, MEMBER_TOKENS.getAccessToken())
+                        .cookie(COOKIE)
+                        .contentType(APPLICATION_JSON)
+                        .content(objectMapper.writeValueAsString(sharedStatusRequest)))
+                .andExpect(status().isOk())
+                .andDo(restDocs.document(
+                                pathParameters(
+                                        parameterWithName("tripId")
+                                                .description("여행 ID")
+                                ),
+                                requestFields(
+                                        fieldWithPath("sharedStatus")
+                                                .type(JsonFieldType.BOOLEAN)
+                                                .description("공유 유무")
+                                                .attributes(field("constraint", "공유시: true, 비공유시: false"))
+                                ),
+                                responseFields(
+                                        fieldWithPath("sharedCode")
+                                                .type(JsonFieldType.STRING)
+                                                .description("공유 코드")
+                                                .attributes(field("constraint", "공유시: 문자열 비공유시: null"))
+                                                .optional()
+                                )
+                        )
+                )
+                .andReturn();
+    }
+
+    @DisplayName("공유 상태가 없는 공유 수정 요청은 예외처리한다.")
+    @Test
+    void getSharedTrip_NullSharedStatus() throws Exception {
+        // given
+        final SharedStatusRequest sharedStatusRequest = new SharedStatusRequest(null);
+        final SharedCodeResponse sharedCodeResponse = new SharedCodeResponse("xxxxxx");
+        when(tripService.updateSharedTripStatus(anyLong(), any(SharedStatusRequest.class)))
+                .thenReturn(sharedCodeResponse);
+        given(refreshTokenRepository.existsByToken(any())).willReturn(true);
+        doNothing().when(jwtProvider).validateTokens(any());
+        given(jwtProvider.getSubject(any())).willReturn("1");
+        doNothing().when(tripService).validateTripByMember(anyLong(), anyLong());
+
+        // when & then
+        mockMvc.perform(patch("/trips/{tripId}/share", 1)
+                        .header(AUTHORIZATION, MEMBER_TOKENS.getAccessToken())
+                        .cookie(COOKIE)
+                        .contentType(APPLICATION_JSON)
+                        .content(objectMapper.writeValueAsString(sharedStatusRequest)))
+                .andExpect(status().isBadRequest())
+                .andExpect(jsonPath("$.message").value("공유 상태를 선택해주세요."));
     }
 
     @DisplayName("커뮤니티 공개 상태를 변경한다")
